If userspace reboot watchdog triggers, don't store reason in persistent property
If init is wedged, then the write will never succeed and reboot won't
happen.

Also, in case of normal reboot, move call to PersistRebootReason to the
top of DoReboot() function, to make sure we persist it even if /data is
not mounted.

Test: builds
Test: adb shell svc power reboot userspace
Test: atest CtsUserspaceRebootHostSideTestCases
Bug: 148767783
Change-Id: I4ae40e1f6fdc41cc0bcae57020fa3d3385dda1b4
Merged-In: I4ae40e1f6fdc41cc0bcae57020fa3d3385dda1b4

Former-commit-id: 9a5df066d578cb2bb651ca11279f39eb69169c43 [formerly dffbb4f14839a4ee6b7f570249f05b8d72fb7319]
Former-commit-id: c8d906f4551c4f54367c22d6bb9c166e691615a7