Whole static link libasync_safe to libbacktrace.a
When libbacktrace.a is statically lined to somewhere, that library had
to add libasync_safe.a to static_libs because libbacktrace.a has
references to libasync_safe.a. But libbacktace depending on
libasync_safe is an implementation detail of libbacktrace, and therefore
its client shouldn't be affected by it.

Fixing this by doing the whole static link to libasync_safe to
libbacktrace.a so that the former is included in libbacktrace.a

Bug: 149569129
Test: m
Change-Id: If7366a240bc945dda9944fe7c111e10d328165bb

Former-commit-id: 6ba069bef669ab65b28eb9f0284078ebde113ddc [formerly 7c8e2b247d2411d76196dc8a56656d39fe40214e]
Former-commit-id: eaa0a2ebfd8f5fc26123dbfc09b5c7d4cf9511a1