don't include liblog to APEXes
liblog is a platform library that provides stable C API. There is no
need to include the library, especialy by statically linking to it, in
any APEX. It not only wastes the storage/ram, but also is incorrect
because the socket interface to logd which is implemented in liblog is
not guaranteed to be stable.

Fixing this issue by converting static_libs: ["liblog"] into
shared_libs: ["liblog"], in which case the dependency to the library
is satisfied via the stub variant of the library.

As a result, we could restrict the availablity of the library to
the platform and the runtime APEX.

Exempt-From-Owner-Approval: already approved when this was in internal
master (ag/10572699)

Bug: http://b/151051671
Bug: http://b/150827719
Test: m
Merged-In: I5aab863cb12b8767b6979255c247000a59355b0e
(cherry picked from commit a1dbcdc2362c96b44f4902abe4bd40cb0f0b1e4a [formerly 95b6f45b0e09b45daf6a7999cef039fabf587327])
Change-Id: I5aab863cb12b8767b6979255c247000a59355b0e

Former-commit-id: 43eb5ff7e29046067af4c4499db03f815768cd9c [formerly 01aa81cec546e625f13375f6d6929cbc77d64965]
Former-commit-id: 27425a2642743fd71d090997fb23bbf6a9ac7ad6