Make libselinux a stub library
libselinux is currently being copied to APEXes. This is risky because
the library is not designed to be portable; part of it is tied to the
specific version of the Android that it was developed for.

This change fixes the problem by declaring that the library supports
a stub with the list of C APIs that are included in the stub. Then there
is only one copy of libselinux in /system/lib and other APEXes use the
copy by dynamically linking to it.

Also, adbd no longer statically links to it, because doing so brings
libselinux in it.

Bug: 151053366
Test: m com.android.adbd. It doesn't include libselinux in it.
Test: m com.android.adbd-deps-info. then inspect
out/soong/com.android.adbd-deps-info.txt. The dependency to libselinux
is shown as '(external)'.

Exempt-From-Owner-Approval: cherry-pick from AOSP

Merged-In: If418cbe3abdeacb759d59052e6dca4c2067678dd
(cherry picked from commit e753ea6c7269440a4712179d9a8da6a063d3ee48 [formerly 3ffdad0cb5bbc3855af7705fee62e2be09583a17])
Change-Id: If418cbe3abdeacb759d59052e6dca4c2067678dd

Former-commit-id: cf7cdef96f1d30a50e78f5211de060fa5d95111b [formerly 8ce4a3067c1f0bdb10262b93f0283b84c5ebe816]
Former-commit-id: 6806061ceaf1aa5c8cf0042a04e10f153f3973a8