Support ERROR_ATOM_ID_INVALID_POSITION
Originally, the native implementation would discard everything clients
wrote to the buffer before the setAtomId call. Thus, the atom id was
always placed in the "correct position." However, for consistency with
the Java implementation, we now allow write calls before setAtomId to
occur and log an error in that case.

Test: bit libstatssocket_test:*
Bug: 152119205
Change-Id: Ib109c7a939a5ae92c1d5fc70aa647ac7390fadad

Former-commit-id: abb0c3ad1df03556e9e030dba29e6ceb516a8991 [formerly 8056dd129a2439e883d9df41cae113c0f0894643]
Former-commit-id: 9c2564efc892343775630c9921af942805cd46c0