Revert "Add "com.android.os.statsd" to libprocessgroup apex_available"
This reverts commit 33e8afc6173ada71f144b9794da59e73f8dc2f7f [formerly 2b859925eaa5ac6ffc04a16f8c6a7a692dfc6db9].

Reason for revert: Not needed

The original change was made in an attempt to close a perceived hole in
the apex availability checking. However, there is no such hole and so
this change is unnecessary. While there is a dependency path from the
apex to this module one of the dependencies in that path is from a
static library to a shared library. That dependency does not result in
the code that includes the static library being linked to the shared
library at either build time or runtime. Therefore, the shared library
is not part of the APEX and so does not need to be marked as such.

Bug: 152762638
Test: m droid
Change-Id: I77c3b23d28eca99fcb133cb38eb8863277ac7179

Former-commit-id: 427a60e5925f50274805479262b309bb738e6f78