Retry pullfinish with empty payload if call fails
If the first call for pullFinished fails, it is likely because the
transaction is too large. Currently, if this happens statsd will just
sleep until the timeout. With this change, the client will retry calling
pullFinish if the first attempt fails, but with an empty payload (and
with success = false) to cause the puller to fail fast, and so statsd
does not wait for the timeout.

Test: atest LibStatsPullTests
Bug: 157768117
Change-Id: I3af991fae47bb7c09bf8a6fc83f28556b836c6e4

Former-commit-id: 775f33fef408197c178c97a5a0a101294ff965d4 [formerly c95f7819017eeeccfb78cab728c3c443546affed]
Former-commit-id: 320959e401fbef0a5484ebba88623d7259acf2b9