fs_mgr: Harden adb-remount-test.sh on read-only filesystem
I'm getting this error:

  [  FAILED  ] /cache/overlay wipe

If /cache is not mounted, in which case it would be RO and empty, then
don't try to do `rm -rf /cache/overlay` as it would fail.
If /cache is RO and /cache/overlay doesn't exist, then `rm -f` would
fail, albeit the "-f" flag, this is a toybox bug.
Check file path existence before `rm`-ing so we don't trigger this error.

Change `ls <type> <path>` to `test <type> <path>` for robust checking.

Bug: 178683776
Bug: 165925766
Bug: 178771232
Test: adb-remount-test.sh on GSI + Pixel
Change-Id: Ie95de690d96146892127ab11a461d80cb6cb56fa

Former-commit-id: 41b617174156c151e4a5dd120d782a2d9bdecb80