Add a core configuration that disables the zygote.
In targets that do not include a zygote binary, attempting to start the
zygote service will result in repeated service restarts. Avoid this by
providing a way for targets to opt out of declaring a zygote service.

The fvp_mini target does not have a zygote, so apply this opt-out to
that target.

Change-Id: I4f918502e611e950fa039e4e2ed817c97b928ba2

Former-commit-id: 1414e3cc7777683c25420f9068d164b5c3163ce8