Update the encryption rule of /data/bootanim to DeleteIfNecessary
Directories should always be encrypted unless there is a specific reason
they can't be.  /data/bootanim is unencrypted without a specific reason,
so fix it to be encrypted.  It is too late to use encryption=Require.
However, the contents of this directory doesn't need to be preserved on
updates, so we can use encryption=DeleteIfNecessary instead of
encryption=Attempt.

Bug: 232299581
Test: build succeeds
Change-Id: Iae80043f3957a053d47ab4754c42b1fdc8e696b3
Merged-In: I17bcb901ad533cada4e0aa061196fc94d7b213ec

Former-commit-id: 86b08e5efdbf8563023a7525f79bae5cfed10da2